+++ pkg/security/password/store.go
@@ -11,9 +11,9 @@ import (
 // NOTE: ownerID represents the ObjectID of whoever owns a given password
 type Store interface {
 	Upsert(ctx context.Context, p Password) error
-	Update(ctx context.Context, k Kind, ownerID int64, newpass []byte) error
-	Get(ctx context.Context, k Kind, ownerID int64) (Password, error)
-	Delete(ctx context.Context, k Kind, ownerID int64) error
+	Update(ctx context.Context, k Kind, ownerID uint32, newpass []byte) error
+	Get(ctx context.Context, k Kind, ownerID uint32) (Password, error)
+	Delete(ctx context.Context, k Kind, ownerID uint32) error
 }
 
 type passwordStore struct {
@@ -56,7 +56,7 @@ func (s *passwordStore) Upsert(ctx context.Context, p Password) (err error) {
 }
 
 // UpdatePolicy updates an existing password record
-func (s *passwordStore) Update(ctx context.Context, k Kind, ownerID int64, newpass []byte) (err error) {
+func (s *passwordStore) Update(ctx context.Context, k Kind, ownerID uint32, newpass []byte) (err error) {
 	if len(newpass) == 0 {
 		return ErrEmptyPassword
 	}
@@ -77,7 +77,7 @@ func (s *passwordStore) Update(ctx context.Context, k Kind, ownerID int64, newpa
 }
 
 // Get retrieves a stored password
-func (s *passwordStore) Get(ctx context.Context, k Kind, userID int64) (p Password, err error) {
+func (s *passwordStore) Get(ctx context.Context, k Kind, userID uint32) (p Password, err error) {
 	// retrieving password
 	err = s.db.NewSession(nil).
 		Select("*").
@@ -97,7 +97,7 @@ func (s *passwordStore) Get(ctx context.Context, k Kind, userID int64) (p Passwo
 }
 
 // DeletePolicy deletes a stored password
-func (s *passwordStore) Delete(ctx context.Context, k Kind, ownerID int64) (err error) {
+func (s *passwordStore) Delete(ctx context.Context, k Kind, ownerID uint32) (err error) {
 	_, err = s.db.NewSession(nil).
 		DeleteFrom("password").
 		Where("kind = ? AND owner_id = ?", k, ownerID).
